requirements.txt
torch
torchvision
Pillow
numpy
flask
flask_cors
adafruit-io
tqdm
